rename: modalPost 컴포넌트명 modalRead로 수정
